Cleanup non used WIF code, removed installed web.config changes back to non install
reverted connectionString to non set value, set version back to empty.
Removed SiteContants SignOutUrl property (not needed).

This should be a clean commit with WIF added, to turn it on set
IsWIFRelyingParty=true in appSettings.  Then setup the sites in IIS and
go through the WIF Setup doc I added to the developer documentation
solution folder.