StorageProviders
The idea is to be able to abstract out how uploaded files are stored and
read in MVCForum. What's included basically leaves everything as-is, but
allows for adding a custom storage provider later on. In our specific
case, a storage provider for Azure was created using this that allowed
storage files in Azure's blob storage, but any number of storage
locations could be used with this. Please let me know if you have any
questions.