Fixes #137
Advantages:
Activity Page almost the same as for VBulletin 4
I hope this improvement is compatible with project Dialogue
Disadvantages:
A heavier EF query (we should do not forget about the category permissions)
Database field "Activity.Data" contains "Value" (post Id or topic Id) instead of "Key=Value". But it looks like this is not a big problem.